+++ pkg/hummingbird/reservation.go
@@ -1,4 +1,4 @@
-// Copyright 2023 ETH Zurich
+// Copyright 2024 ETH Zurich
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
